requirements.txt
Django==4.2.7
djangorestframework==3.14.0
numpy==1.25.2
scikit-learn==1.3.1
openai==0.28.1